+++ R/bs-theme.R
@@ -120,10 +120,7 @@ bs_theme <- function(version = version_default(), bootswatch = NULL, ...,
   theme <- bs_bundle(
     bs_theme_init(version, bootswatch),
     bootstrap_bundle(version),
-    bootswatch_bundle(bootswatch, version),
-    # Always include color-contrast() so we can support bs_get_contrast()
-    # for any version, and so 3rd party bslib extensions can always use it
-    color_contrast_layer()
+    bootswatch_bundle(bootswatch, version)
   )
   bs_theme_update(
     theme, ...,
@@ -160,7 +157,7 @@ bs_theme_update <- function(theme, ..., bootswatch = NULL, bg = NULL, fg = NULL,
     }
     if (!identical(bootswatch, "default")) {
       theme <- add_class(theme, bootswatch_class(bootswatch))
-      theme <- bs_bundle(theme, bootswatch_bundle(bootswatch, theme_version(theme)), color_contrast_layer())
+      theme <- bs_bundle(theme, bootswatch_bundle(bootswatch, theme_version(theme)))
     }
   }
   # See R/bs-theme-update.R for the implementation of these
@@ -233,17 +230,6 @@ assert_bs_theme <- function(theme) {
   }
   invisible(theme)
 }
-# -----------------------------------------------------------------
-# Color contrast layer
-# -----------------------------------------------------------------
-
-color_contrast_layer <- function() {
-  sass_layer(
-    defaults = sass_file(
-      system_file("sass-utils/color-contrast.scss", package = "bslib")
-    )
-  )
-}
 
 # -----------------------------------------------------------------
 # Core Bootstrap bundle
@@ -258,18 +244,20 @@ bootstrap_bundle <- function(version) {
     ".table th[align=center] { text-align: center; }"
   )
 
-  res <- switch_version(
+  main_bundle <- switch_version(
     version,
     five = sass_bundle(
       # Don't name this "core" bundle so it can't easily be removed
       sass_layer(
-        defaults = bs5_sass_files(c("functions", "variables")),
-        declarations = bs5_sass_files(c("mixins", "utilities"))
+        functions = bs5_sass_files("functions"),
+        defaults = bs5_sass_files("variables"),
+        mixins = bs5_sass_files("mixins")
       ),
       # Returns a _named_ list of bundles (i.e., these should be easily removed)
       !!!rule_bundles(
         # Names here should match https://github.com/twbs/bs5/blob/master/scss/bootstrap.scss
         bs5_sass_files(c(
+          "utilities",
           "root", "reboot", "type", "images", "containers", "grid",
           "tables", "forms", "buttons", "transitions", "dropdown",
           "button-group", "nav", "navbar", "card", "accordion", "breadcrumb",
@@ -284,7 +272,8 @@ bootstrap_bundle <- function(version) {
     ),
     four = sass_bundle(
       sass_layer(
-        defaults = bs4_sass_files(c("deprecated", "functions", "variables")),
+        functions = bs4_sass_files(c("deprecated", "functions")),
+        defaults = bs4_sass_files("variables"),
         mixins = bs4_sass_files("mixins")
       ),
       # Returns a _named_ list of bundles (i.e., these should be easily removed)
@@ -328,12 +317,22 @@ bootstrap_bundle <- function(version) {
       )
     )
   )
-  # Tack on nav_spacer() CSS by default
-  nav_spacer <- sass_file(
-    system.file("nav-spacer", "nav-spacer.scss", package = "bslib")
-  )
 
-  sass_bundle(res, nav_spacer = sass_layer(rules = nav_spacer))
+  sass_bundle(
+    main_bundle,
+    # color-contrast() was introduced in Bootstrap 5.
+    # We include our own version for a few reasons:
+    # 1. Easily turn off warnings options(bslib.color_contrast_warnings=F)
+    # 2. Allow Bootstrap 3 & 4 to use color-contrast() in variable definitions
+    # 3. Allow Bootstrap 3 & 4 to use bs_get_contrast()
+    sass_layer(
+      functions = sass_file(system_file("sass-utils/color-contrast.scss", package = "bslib"))
+    ),
+    # nav_spacer() CSS (can be removed)
+    nav_spacer = sass_layer(
+      rules = sass_file(system_file("nav-spacer/nav-spacer.scss", package = "bslib"))
+    )
+  )
 }